Can't forget the 1982!
The Commodore 64, also known as the C64 or the CBM 64, is an 8-bit home computer introduced in January 1982 by Commodore International (first shown at the Consumer Electronics Show, January 7–10, 1982, in Las Vegas).[4] It has been listed in the Guinness World Records as the highest-selling single computer model of all time,[5] with independent estimates placing the number sold between 12.5 and 17 million units.[2] Volume production started in early 1982, marketing in August for US$595 (equivalent to $1,596 in 2020).[6][7] Preceded by the Commodore VIC-20 and Commodore PET, the C64 took its name from its 64 kibibytes (65,536 bytes) of RAM. With support for multicolor sprites and a custom chip for waveform generation, the C64 could create superior visuals and audio compared to systems without such custom hardware.

The C64 dominated the low-end computer market (except in the UK and Japan, lasting only about six months in Japan[8]) for most of the later years of the 1980s.[9] For a substantial period (1983–1986), the C64 had between 30% and 40% share of the US market and two million units sold per year,[10] outselling IBM PC compatibles, Apple computers, and the Atari 8-bit family of computers. Sam Tramiel, a later Atari president and the son of Commodore's founder, said in a 1989 interview, "When I was at Commodore we were building 400,000 C64s a month for a couple of years."[11] In the UK market, the C64 faced competition from the BBC Micro and the ZX Spectrum,[12] but the C64 was still the second most popular computer in the UK after the ZX Spectrum.[13] The Commodore 64 failed to make any impact in Japan. The Japanese market was dominated by Japanese computers, such as the NEC PC-8801, Sharp X1, Fujitsu FM-7, and MSX.[14]

Part of the Commodore 64's success was its sale in regular retail stores instead of only electronics or computer hobbyist specialty stores. Commodore produced many of its parts in-house to control costs, including custom integrated circuit chips from MOS Technology. In the United States, it has been compared to the Ford Model T automobile for its role in bringing a new technology to middle-class households via creative and affordable mass-production.[15] Approximately 10,000 commercial software titles have been made for the Commodore 64, including development tools, office productivity applications, and video games.[16] C64 emulators allow anyone with a modern computer, or a compatible video game console, to run these programs today. The C64 is also credited with popularizing the computer demoscene and is still used today by some computer hobbyists.[17] In 2011, 17 years after it was taken off the market, research showed that brand recognition for the model was still at 87%.[5]